refactor(functions): extract move to param
