+++ test/functions_test.js
@@ -1,7 +1,12 @@
 // @ts-check
 
 import { strict as assert } from "assert";
-import { passToParam, resignToParam, swapToParam } from "../src/functions.js";
+import {
+  moveToParam,
+  passToParam,
+  resignToParam,
+  swapToParam,
+} from "../src/functions.js";
 
 {
   /** @type {import("../src/models").Pass} */
@@ -32,4 +37,19 @@ import { passToParam, resignToParam, swapToParam } from "../src/functions.js";
   };
   const expected = "1|s|a|b";
   assert.equal(swapToParam(input), expected);
+}
+
+{
+  /** @type {import("../src/models").Move} */
+  const input = {
+    type: "move",
+    playerId: 0,
+    coordinates: [
+      { panel: "a", x: 0, y: 0 },
+      { panel: "r", x: 1, y: 0 },
+      { panel: "m", x: 2, y: 0 },
+    ],
+  };
+  const expected = "0|m|0|0|a|1|0|r|2|0|m";
+  assert.equal(moveToParam(input), expected);
 }
